Move step model in a module exposing a factory.
