+++ lib/parser/featureModels/scenarioOrBackgroundFactory.js
@@ -1,12 +1,7 @@
 'use strict';
 /* eslint camelcase: 0 */
 
-function Step(config) {
-  this.token = 'step';
-  this.keyword = config.keyword;
-  this.name = config.name;
-  this.line = config.line;
-}
+var stepFactory = require('./stepFactory');
 
 function ScenarioOrBackground(config) {
   this.steps = [];
@@ -20,7 +15,7 @@ function ScenarioOrBackground(config) {
 }
 
 ScenarioOrBackground.prototype.addStep = function(config) {
-  this.steps.push(new Step(config));
+  this.steps.push(stepFactory(config));
 }
 
 module.exports = function(config) {
